Fix target and targetname base classes; apply property type hot fix
When the first part of the property type hot fix was merged, the `target` property in the Target base class was accidentally set to `target_source` instead of `target_destination`, and vice versa for the `targetname` property in the Targetname base class. They have been switched to their correct property types.

The recent property type hot fix for the plugin has also been applied.